+++ src/features/review-filtering/lib/isReviewCategory.ts
@@ -1,5 +1,5 @@
-import {ReviewCategory} from '@/entities/reviews';
 import CATEGORY_LIST from '../consts/categoryList';
+import {ReviewCategory} from '@/entities/reviews';
 
 export function IsReviewCategory(value: string | null): value is ReviewCategory {
   return value !== null && CATEGORY_LIST.some(({id}) => id === value);
